Split configurator class in two.
Configurator purely loads up the config file, and urls file now. The
base class handles the connection to the database and updating the
podcasts list.

This allows me to call configurator on it's own without redoing the
time consuming process of importing the podcasts each time. As a
result commands/* now inherit the base class, which inherts the
configurator class.

Mixins are now in their own Mixins module.

The downloader class can now inhert from the Configurator class without a
performance hit.